docs(headless): improve automatic facet doc to reflect what should be used (#3318)
https://coveord.atlassian.net/browse/KIT-2838

* better comments for automatic facet props and buildautomaticFAcet

https://coveord.atlassian.net/browse/KIT-2838

* better comments for automatic facet props and buildautomaticFAcet

https://coveord.atlassian.net/browse/KIT-2838

* remove the headless automatic facet page

https://coveord.atlassian.net/browse/KIT-2838

* Update packages/headless/src/controllers/facets/automatic-facet/headless-automatic-facet.ts


* Update packages/headless/src/controllers/facets/automatic-facet/headless-automatic-facet.ts


* remove comment

https://coveord.atlassian.net/browse/KIT-2838

* Update packages/headless/src/controllers/facets/automatic-facet-generator/headless-automatic-facet-generator.ts


---------
