+++ packages/headless/src/controllers/facets/automatic-facet-generator/headless-automatic-facet-generator.ts
@@ -14,23 +14,22 @@ import {
   Controller,
   buildController,
 } from '../../controller/headless-controller';
-import {
-  AutomaticFacet,
-  buildAutomaticFacet,
-} from '../automatic-facet/headless-automatic-facet';
+import {buildAutomaticFacet} from '../automatic-facet/headless-automatic-facet';
+import {FacetValue} from '../facet/headless-facet';
 import {
   AutomaticFacetGeneratorOptions,
   buildOptions,
 } from './headless-automatic-facet-generator-options';
 
 export type {AutomaticFacetGeneratorOptions};
+
 /**
  * The `AutomaticFacetGenerator` headless controller offers a high-level interface for rendering automatic facets.
  *
- * Unlike traditional facets that need to be explicitly defined and requested in the query, automatic facets are dynamically generated by the index
- * in response to the query.
+ * Unlike regular facets that need to be explicitly defined and requested in the query, automatic facets are dynamically
+ * generated by the index in response to the query.
  *
- * To read more about the automatic facet generator feature, see: https://docs.coveo.com/en/n9sd0159/
+ * To learn more about the automatic facet generator feature, see: [About the Facet Generator](https://docs.coveo.com/en/n9sd0159/).
  */
 export interface AutomaticFacetGenerator extends Controller {
   /**
@@ -51,6 +50,48 @@ export interface AutomaticFacetGeneratorProps {
    * The options for the `AutomaticFacetGenerator` controller.
    * */
   options: AutomaticFacetGeneratorOptions;
+}
+
+/**
+ * A scoped and simplified part of the headless state that is relevant to the `Automatic Facet` controller.
+ */
+export interface AutomaticFacetState {
+  /**
+   * The automatic facet field.
+   */
+  field: string;
+  /**
+   * The automatic facet label.
+   */
+  label: string;
+  /**
+   * The automatic facet values.
+   */
+  values: FacetValue[];
+}
+/**
+ * The `AutomaticFacet` controller allows you to create a search interface component that the end user
+ * can interact with to refine a query by selecting filters based on item metadata (i.e., field values). Unlike regular facets that
+ * need to be explicitly defined and requested in the query, automatic facets are dynamically generated by the index
+ * in response to the query.
+ *
+ * To learn more about the automatic facet generator feature, see: [About the Facet Generator](https://docs.coveo.com/en/n9sd0159/).
+ */
+export interface AutomaticFacet extends Controller {
+  /**
+   * Toggles the specified automatic facet value.
+   *
+   * @param selection - The facet value to toggle.
+   */
+  toggleSelect(selection: FacetValue): void;
+  /**
+   * Deselects all automatic facet values.
+   * */
+  deselectAll(): void;
+  /**
+   *  The state of the `AutomaticFacet` controller.
+   */
+  state: AutomaticFacetState;
 }
 
 /**
