+++ packages/headless/src/controllers/facets/automatic-facet/headless-automatic-facet.ssr.ts
@@ -1,7 +1,7 @@
 import {SearchEngine} from '../../../app/search-engine/search-engine';
 import {ControllerDefinitionWithoutProps} from '../../../app/ssr-engine/types/common';
+import {AutomaticFacet} from '../automatic-facet-generator/headless-automatic-facet-generator';
 import {
-  AutomaticFacet,
   AutomaticFacetProps,
   buildAutomaticFacet,
 } from './headless-automatic-facet';
