+++ packages/headless/src/controllers/facets/automatic-facet/headless-automatic-facet.ts
@@ -7,30 +7,12 @@ import {logFacetClearAll} from '../../../features/facets/facet-set/facet-set-ana
 import {getAnalyticsActionForToggleFacetSelect} from '../../../features/facets/facet-set/facet-set-utils';
 import {FacetValue} from '../../../features/facets/facet-set/interfaces/response';
 import {executeSearch} from '../../../features/search/search-actions';
-import {
-  Controller,
-  buildController,
-} from '../../controller/headless-controller';
+import {buildController} from '../../controller/headless-controller';
+import {AutomaticFacet} from '../automatic-facet-generator/headless-automatic-facet-generator';
 
 /**
- * A scoped and simplified part of the headless state that is relevant to the `Automatic Facet` controller.
- */
-export interface AutomaticFacetState {
-  /**
-   * The automatic facet field.
-   */
-  field: string;
-  /**
-   * The automatic facet label.
-   */
-  label: string;
-  /**
-   * The automatic facet values.
-   */
-  values: FacetValue[];
-}
-/**
- * The props for the `AutomaticFacet` controller.
+ * @internal
+ * This prop is used internally by the `AutomaticFacetGenerator` controller.
  */
 export interface AutomaticFacetProps {
   /**
@@ -38,39 +20,16 @@ export interface AutomaticFacetProps {
    */
   field: string;
 }
+
 /**
- * The `AutomaticFacet` controller allows you to create a search interface component that the end user
- * can use to refine a query by selecting filters based on item metadata (i.e., field values). Unlike traditional facets that
- * need to be explicitly defined and requested in the query, automatic facets are dynamically generated by the index
- * in response to the query.
+ * @internal
+ * This initializer is used internally by the `AutomaticFacetGenerator` controller.
  *
- * It's important to note that this controller is not typically used directly.
- * It is internally used by the `AutomaticFacetGenerator` controller to automatically render updated facets.
- *
- * To read more about the automatic facet generator feature, see: https://docs.coveo.com/en/n9sd0159/
- */
-export interface AutomaticFacet extends Controller {
-  /**
-   * Toggles the specified automatic facet value.
-   *
-   * @param selection - The facet value to toggle.
-   */
-  toggleSelect(selection: FacetValue): void;
-  /**
-   * Deselects all automatic facet values.
-   * */
-  deselectAll(): void;
-  /**
-   *  The state of the `AutomaticFacet` controller.
-   */
-  state: AutomaticFacetState;
-}
-/**
- * Creates a `AutomaticFacet` controller instance.
+ * **Important:** This initializer is meant for internal use by Headless only. As an implementer, you should never import or use this initializer directly in your code.
  *
  * @param engine - The headless engine.
- * @param props - The configurable `AutomaticFacet` properties.
- * @returns A `AutomaticFacet` controller instance.
+ * @param props - The configurable `AutomaticFacet` properties used internally.
+ * @returns An `AutomaticFacet` controller instance.
  * */
 export function buildAutomaticFacet(
   engine: SearchEngine,
